+++ game_ctrl.cpp
@@ -242,12 +242,12 @@ CControl *CPlayers::GetCtrl () {
 	return plyr[currPlayer].ctrl;
 }
 
-CControl *CPlayers::GetCtrl (int player) {
+CControl *CPlayers::GetCtrl (size_t player) {
 	if (player < 0 || player >= plyr.size()) return NULL;
 	return plyr[player].ctrl;
 }
 
-string CPlayers::GetName (int player) const {
+string CPlayers::GetName (size_t player) const {
 	if (player < 0 || player >= plyr.size()) return "";
 	return plyr[player].name;
 }
@@ -262,7 +262,7 @@ void CPlayers::ResetControls () {
 }
 
 // called in module regist.cpp:
-void CPlayers::AllocControl (int player) {
+void CPlayers::AllocControl (size_t player) {
 	if (player < 0 || player >= plyr.size()) return;
 	if (plyr[player].ctrl != NULL) return;
 	plyr[player].ctrl = new (CControl);
